+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Define configs in modal Extensions.
 * Connect Discord via RPC.
+* Documentation from installing
 * Change status microphone (enable, disable, toggle).
 * Change status Deaf (enable, disable, toggle)
-* Documentation from installing
-* 
+* Disconnect voice chat